Solution to Leet Code problem https://leetcode.com/problems/minimum-moves-to-equal-array-elements-ii/?tab=Description
First attempt: Put all elements in a map and then iterate through the map checking if each element is less than or greater than mid value.  If so, increment amount of moves required by the difference between the mid value and current valued, multiplied by number of items in the index if there is more than one of that element.