Solution to Leet Code problem https://leetcode.com/problems/find-all-numbers-disappeared-in-an-array/?tab=Description
Second attempt: Iterate through the array indices and take absolute value of the element minus 1 (valueSeen) to give us the index BEFORE the element we are recording as "seen".  If the element at the array index before the value we've now seen is greater than 0 (not negative), set that element to negative.

Then, iterate once more through the array and any values at array[i]  + 1 that are not negative, do not exist in the array