Solution to Leet Code problem 'Roman Numeral to Integer' https://leetcode.com/problems/roman-to-integer/#/description
Assign all roman numerals to a map with the corresponding int value.  Iterate through the string as a char array and add to the result value.  Check if the current value, in the iteration, is greater than the last value you used.  If it is, subtract the last value and then add (current numeral value - last numeral value ) to the result